backend/requirements.txt
fastapi
uvicorn
langchain
pydantic
python-dotenv
langchain-groq
langchain-exa
exa-py
datasets
pandas
scikit-learn
tqdm
scipy
